# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "calm-server",
    "description": "Wire-format contract between calm-server (Rust) and web-calm (TS). Source of truth for generated TypeScript types.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/waves/{wave_id}/terminal-cards": {
      "post": {
        "tags": [
          "terminals"
        ],
        "operationId": "create_terminal_card",
        "parameters": [
          {
            "name": "wave_id",
            "in": "path",
            "description": "Wave id to create the terminal card under",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Body required (theme is mandatory; program/cwd/env optional)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTerminalCardBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Card + linked terminal created atomically; daemon spawned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Wave not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Body missing required fields (e.g. theme)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Daemon spawn failed; the saga rolled back the committed transaction (no leaked rows).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentProvider": {
        "type": "string",
        "enum": [
          "codex",
          "claude"
        ]
      },
      "Card": {
        "type": "object",
        "required": [
          "id",
          "wave_id",
          "kind",
          "sort",
          "payload",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "deletable": {
            "type": "boolean",
            "description": "Issue #229 PR A — system-card guard. `true` for user-facing cards\n(the default; all pre-#229 rows backfill via the column DEFAULT in\nmigration 0013). `false` for kernel-owned cards that the user\ncannot remove via REST / plugin callbacks — currently spec cards\n(retroactively undeletable via the same migration's UPDATE) and\nPR B's wave-report cards.\n\n`#[serde(default = \"default_deletable\")]` so wire payloads emitted\nbefore #229 landed (event-log replay fixtures, old test seeds)\nparse as `true` without forcing a fixture rewrite — matches the\nDB DEFAULT (1) in migration 0013. The default-fn lives below\nbecause `bool::default()` would give `false` (the *un*safe\nfallback for a deny-by-omission auth bit)."
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "description": "`\"terminal\"` for built-in PTY cards, `\"ui://<plugin>/<view>\"` for\nplugin-provided cards (the canonical MCP Apps resource URI). The\nkernel never interprets beyond that prefix. `[legacy]`\n`\"plugin:<plugin-id>:<view-id>\"` may also appear on rows persisted\nbefore the M4 cut-over and in server-side perms/manifest enforcement\n— see `docs/architecture/terminology-glossary.md` (plugin card kind)."
          },
          "payload": {
            "type": "object",
            "description": "Opaque JSON blob — ts-rs would otherwise emit `unknown` via the\n`serde-json-impl` feature, but we pin it explicitly so a future\nfeature-flag change can't silently widen / narrow the surface."
          },
          "runtime": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CardRuntimeView"
              }
            ]
          },
          "sort": {
            "type": "number",
            "format": "double"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          },
          "wave_id": {
            "type": "string"
          }
        }
      },
...
      "CardRuntimeView": {
        "type": "object",
        "description": "Live runtime projection joined from the `runtimes` table when a card is\nfetched or serialized.\n\nThis view is not part of the idempotency contract: across retries the\nruntime row may have advanced, so `Card.runtime` may differ between the\nfirst POST response and a retry POST response returning the same operation\nresult. Future cleanup (#581 item 4) will remove the legacy payload-key\nprojection; this typed view is the forward-compatible reader path.",
        "required": [
          "runtime_id",
          "kind",
          "status"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/WorkerSessionKind"
          },
          "provider": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentProvider"
              }
            ]
          },
          "runtime_id": {
            "type": "string"
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/WorkerSessionState"
          },
          "terminal_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "thread_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "thread_status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "JSON shape returned for every error response — `{error, code}`.\nMirrors the body produced by `CalmError::into_response`. Hand-written\nduplicate of the in-line `json!` body so OpenAPI consumers see a\nconcrete schema.",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code — one of `not_found`, `conflict`,\n`idempotency_collision`, `bad_request`, `unauthorized`,\n`forbidden`, `plugin_install`, `plugin_permission`,\n`plugin_conflict`, `plugin_kernel_too_old`,\n`spec_harness_dormant`, `db_error`, `io_error`, `serde_error`,\n`codex_app_server`, `service_unavailable`, `internal`,\n`forbidden_tool`, `not_a_card_tool`, `tool_call_failed`."
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "NewTerminalCardBody": {
        "type": "object",
        "description": "Body for `POST /api/waves/:wave_id/terminal-cards`.\n\nDeliberately omits `kind` (always `\"terminal\"`) and `payload` (the kernel\npersists schema payload and projects identity from `runtimes`). Empty\n`program` falls back to `$SHELL` then `/bin/sh`; empty `cwd` falls back to\n`$HOME` then the server's cwd. `env` is merged into the daemon's environment\nas additional vars on top of `TERM` / `COLORTERM` / inherited.",
        "required": [
          "theme"
        ],
        "properties": {
          "cwd": {
            "type": "string",
            "description": "Empty string or missing → `$HOME` (then cwd of server)."
          },
          "env": {
            "type": "object",
            "description": "Extra env on top of the inherited set. JSON object: `{\"FOO\":\"bar\"}`."
          },
          "program": {
            "type": "string",
            "description": "Empty string or missing → `$SHELL` (then `/bin/sh`)."
          },
          "sort": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Sort order within the wave. `None` defaults to \"append to end\"."
          },
          "theme": {
            "$ref": "#/components/schemas/RequestTheme",
            "description": "Host browser's current theme RGB (#177). Required — the kernel\nwrites it onto the terminal row inside the same transaction\nthat mints the card, and every spawn for this row reads\n`term.theme_fg/_bg` to stamp `--terminal-fg/-bg` daemon argv."
          }
        }
      },
...
      "RequestTheme": {
        "type": "object",
        "description": "Wire shape of `NewCodexCardBody.theme` / `NewWave.theme`. Matches the\n`calm_session::TerminalTheme` value type one-for-one — duplicated\nhere so the route can keep its own `ToSchema` derive (the\n`calm_session` crate is utoipa-free).",
        "required": [
          "fg",
          "bg"
        ],
        "properties": {
          "bg": {
            "type": "array",
            "items": false,
            "prefixItems": [
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            ]
          },
          "fg": {
            "type": "array",
            "items": false,
            "prefixItems": [
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "WorkerSessionKind": {
        "type": "string",
        "enum": [
          "terminal",
          "codex",
          "claude",
          "shared-spec"
        ]
      },
      "WorkerSessionState": {
        "type": "string",
        "description": "Session state machine column (`worker_sessions.state`, issue #679 §1).\nSingle runtime/session state vocabulary (`worker_sessions.state`, issue #679 §1).",
        "enum": [
          "starting",
          "running",
          "idle",
          "turn_pending",
          "exited",
          "failed",
          "superseded"
        ]
      }
    }
  },
...
}